requirements.txt
PySide6
yara-python
cx_Freeze
#py7zr 
#requests
